Extract the header from the homepage
