+++ public/pages/index.tsx
@@ -1,5 +1,4 @@
-import type { FunctionalComponent, VNode } from 'preact'
-import { useCallback } from 'preact/hooks'
+import type { FunctionalComponent } from 'preact'
 import catsUrl from 'url:../images/cats.png'
 import componentsUrl from 'url:../images/components.png'
 import computerUrl from 'url:../images/computer.png'
@@ -10,36 +9,12 @@ import thumbsUrl from 'url:../images/thumbs.png'
 import { Layout } from '../components/layout'
 import { categories, categoryToSlugs } from '../resources'
 import styles from './index.module.css'
+import { Header } from './index/header'
 
 export const Index: FunctionalComponent = () => {
-  const click = useCallback((e: MouseEvent) => {
-    e.preventDefault()
-    // NOTE: we stop prop so the router doesn't grab it and pushState
-    e.stopPropagation()
-    console.debug('clicked fav button')
-  }, [])
-
   return (
     <Layout>
-      <header class={styles.header}>
-        <div class={styles.headerTitle}>
-          <h1 class={styles.headerTitleHeading}>
-            <abbr title='Progressive Web App'>PWA</abbr> Resources
-          </h1>
-          <p class={styles.headerTitleSubheading}>A curated collection</p>
-        </div>
-        <aside class={styles.headerAside}>
-          <p class={styles.headerAsideText}>Progressive Web Apps</p>
-          <p class={styles.headerAsideButtonWrapper}>
-            <a href='/favs' class={['button', styles.headerAsideButton].join(' ')} onClick={click}>
-              View your <abbr title='favorites'>♥‘s</abbr>
-            </a>
-          </p>
-          <figure class={styles.headerAsideFigure}>
-            <img src={thumbsUrl} width='462' height='360' />
-          </figure>
-        </aside>
-      </header>
+      <Header />
       <hr />
       <main>
         <div>
